Struct numbers::Matrix

pub struct Matrix<T: Num + Clone + Rand> {
    pub elements: Vec<T>,
    pub row_size: usize,
    pub col_size: usize,
    pub transpose: bool,
}

Generic Matrix object.

Fields

elements

row_size

col_size

transpose

Methods

impl<T: Num + Clone + Rand> Matrix<T>

fn new(e: Vec<T>, r_size: usize, c_size: usize) -> Result<Matrix<T>, MatrixError>

Create a new matrix

fn zeros(r_size: usize, c_size: usize) -> Matrix<T>

Create a matrix of zeros

fn random(r_size: usize, c_size: usize) -> Matrix<T>

fn diag_mat(a: Vec<T>) -> Matrix<T>

Create a square diagonal matrix from a vector of elements.

fn identity(row_size: usize) -> Matrix<T>

Create an square identity matrix of specified row size.

fn replace(&mut self, row: usize, col: usize, value: T)

Replace a value in the matrix at specified row and column indices.

fn get_ind(&self, row: usize, col: usize) -> usize

Map an index in a matrix to index in its corresponding 1-d vector.

fn get_element(&self, row: usize, col: usize) -> T

Get an element from the matrix.

fn transpose(&self) -> Matrix<T>

Transpose the matrix. We're not actually changing anything in memory;

fn diagonal(&self) -> Vec<T>

Get the diagonal of a matrix.

fn tri(row_size: usize, col_size: usize, k: usize, upper_or_lower: Triangular) -> Matrix<T>

Get a triangular matrix with specified dimensions above or below specified diagonal.

fn submatrix(&self, start: (usize, usize), dim: (usize, usize)) -> Matrix<T>